feat:[AI-BOT] IMage generation
