+++ python/image-generation/generate-stable-fusion.py
@@ -1,0 +1,336 @@
+import logging
+import sys
+import os
+from datetime import datetime
+import torch
+import subprocess
+import gc
+import json
+from pathlib import Path
+from diffusers import StableDiffusionPipeline
+
+# Set PyTorch memory allocation configuration
+os.environ["PYTORCH_CUDA_ALLOC_CONF"] = "max_split_size_mb:512,expandable_segments:True"
+
+# Configure logging
+logging.basicConfig(
+    level=logging.INFO,
+    format='%(asctime)s - %(levelname)s - %(message)s',
+    handlers=[
+        logging.StreamHandler(sys.stdout),
+        logging.FileHandler('image_generation.log')
+    ]
+)
+logger = logging.getLogger(__name__)
+
+def check_dependencies():
+    """Check and install required dependencies."""
+    try:
+        logger.info("Installing PyTorch...")
+        subprocess.check_call([sys.executable, "-m", "pip", "install", "--no-cache-dir", "torch", "torchvision", "torchaudio", "--index-url", "https://download.pytorch.org/whl/cu118"])
+
+        logger.info("Installing other dependencies...")
+        required_packages = [
+            "diffusers[torch]>=0.21.0",
+            "transformers>=4.25.1",
+            "accelerate>=0.21.0",
+            "safetensors>=0.3.1"
+        ]
+
+        for package in required_packages:
+            try:
+                subprocess.check_call([sys.executable, "-m", "pip", "install", "--no-cache-dir", package])
+            except subprocess.CalledProcessError as e:
+                logger.error(f"Failed to install {package}: {str(e)}")
+                raise
+
+        # Verify imports
+        import diffusers
+        import transformers
+        import accelerate
+        logger.info("All dependencies installed successfully")
+        return True
+
+    except Exception as e:
+        logger.error(f"Error checking/installing dependencies: {str(e)}")
+        raise
+
+def clear_gpu_memory():
+    """Clear GPU memory and run garbage collection."""
+    if torch.cuda.is_available():
+        torch.cuda.empty_cache()
+        torch.cuda.ipc_collect()
+    gc.collect()
+    logger.info("GPU memory cleared")
+
+def setup_pipeline(model_id: str):
+    """Setup the SDXL pipeline with optimizations for limited VRAM."""
+    import torch
+    from diffusers import StableDiffusionPipeline  # Using standard SD instead of SDXL for lower VRAM
+
+    logger.info("Setting up pipeline...")
+    try:
+        # Check CUDA availability
+        if not torch.cuda.is_available():
+            raise RuntimeError("CUDA is not available")
+
+        # Clear GPU memory before loading model
+        clear_gpu_memory()
+
+        # Get available VRAM and set limit before model loading
+        gpu_memory = torch.cuda.get_device_properties(0).total_memory / 1024**3
+        logger.info(f"Available GPU memory: {gpu_memory:.2f} GB")
+
+        # Set memory efficient settings
+        torch.backends.cudnn.benchmark = False  # Disable for stability
+        torch.backends.cuda.matmul.allow_tf32 = True
+        torch.backends.cudnn.allow_tf32 = True
+
+        # Load pipeline with optimizations
+        pipe = StableDiffusionPipeline.from_pretrained(
+            "runwayml/stable-diffusion-v1-5",  # Using SD 1.5 instead of SDXL
+            torch_dtype=torch.float16,
+            safety_checker=None,
+            requires_safety_checker=False
+        )
+
+        # Memory optimizations
+        pipe.enable_attention_slicing()
+        pipe = pipe.to("cuda")
+
+        logger.info("Pipeline setup completed successfully")
+        return pipe
+
+    except Exception as e:
+        logger.error(f"Error setting up pipeline: {str(e)}")
+        raise
+
+def generate_image(
+    pipe,
+    prompt: str,
+    negative_prompt: str,
+    output_file: str,
+    num_steps: int = 30,
+    guidance_scale: float = 7.5,
+    size: tuple = (512, 512)  # Reduced size
+):
+    """Generate image with error handling and progress logging."""
+    try:
+        logger.info("Starting image generation...")
+        logger.info(f"Prompt: {prompt}")
+        logger.info(f"Settings: {num_steps} steps, {guidance_scale} guidance scale, {size} resolution")
+
+        # Clear GPU memory before generation
+        clear_gpu_memory()
+
+        # Generate image
+        with torch.inference_mode():  # Removed autocast for stability
+            result = pipe(
+                prompt=prompt,
+                negative_prompt=negative_prompt,
+                num_inference_steps=num_steps,
+                guidance_scale=guidance_scale,
+                height=size[0],
+                width=size[1]
+            )
+
+            if not result.images:
+                raise RuntimeError("No image was generated")
+
+            image = result.images[0]
+
+        # Clear GPU memory after generation
+        clear_gpu_memory()
+
+        # Create output directory if it doesn't exist
+        output_dir = os.path.dirname(output_file)
+        if output_dir:
+            os.makedirs(output_dir, exist_ok=True)
+
+        # Save with timestamp
+        timestamp = datetime.now().strftime("%Y%m%d_%H%M%S")
+        filename, ext = os.path.splitext(output_file)
+        output_path = f"{filename}_{timestamp}{ext}"
+
+        image.save(output_path)
+        logger.info(f"Image saved successfully to {output_path}")
+
+        return output_path
+
+    except Exception as e:
+        logger.error(f"Error generating image: {str(e)}")
+        raise
+
+def load_response_data(json_file: str):
+    """Load and parse response data from JSON file."""
+    try:
+        with open(json_file, 'r') as f:
+            data = json.load(f)
+        return data
+    except Exception as e:
+        logger.error(f"Error loading response data: {str(e)}")
+        raise
+
+def count_tokens(text: str) -> int:
+    """Roughly estimate token count - this is a simple approximation."""
+    # Split on spaces and punctuation
+    words = text.replace('.', ' . ').replace(',', ' , ').replace('!', ' ! ').replace('?', ' ? ').split()
+    # Count each word/punctuation as a token
+    return len(words)
+
+def create_prompt_from_response(response: str) -> str:
+    """Create a focused prompt directly from the response text."""
+    # Standard suffix we'll always add
+    suffix = " Create a beautiful, detailed landscape. Trending on ArtStation, highly detailed, sharp focus, 8k resolution, cinematic lighting, professional photography, masterpiece"
+    suffix_tokens = count_tokens(suffix)
+
+    # Take sentences until we hit the token limit
+    sentences = response.split('.')
+    current_text = ""
+    total_tokens = 0
+
+    for sentence in sentences:
+        if not sentence.strip():
+            continue
+
+        # Check if adding this sentence would exceed the limit
+        test_text = current_text + sentence + "." if current_text else sentence + "."
+        test_tokens = count_tokens(test_text) + suffix_tokens
+
+        if test_tokens > 70:  # Using 70 as a safety margin
+            break
+
+        current_text = test_text
+
+    # If we couldn't get any sentences under the limit, take just the first one
+    # and truncate it to fit
+    if not current_text:
+        first_sentence = sentences[0] + "."
+        words = first_sentence.split()
+        current_text = " ".join(words[:50])  # Arbitrary limit that should fit
+
+    # Create the final prompt
+    prompt = f"{current_text.strip()}"
+    prompt += suffix
+
+    return prompt
+
+def batch_generate_images(response_dir: str, output_dir: str = "outputs/generated_images"):
+    """Generate images for all JSON files in the response directory."""
+    try:
+        # Create output directory
+        os.makedirs(output_dir, exist_ok=True)
+
+        # Setup pipeline once for all generations
+        pipe = setup_pipeline("runwayml/stable-diffusion-v1-5")
+
+        # Get all JSON files in the directory
+        json_files = sorted([f for f in os.listdir(response_dir) if f.endswith('.json')])
+        total_files = len(json_files)
+        logger.info(f"Found {total_files} JSON files to process")
+
+        for file_idx, json_file in enumerate(json_files, 1):
+            try:
+                logger.info(f"\n{'='*50}")
+                logger.info(f"Processing file {file_idx}/{total_files}: {json_file}")
+                logger.info(f"{'='*50}")
+
+                # Load JSON data
+                file_path = os.path.join(response_dir, json_file)
+                data = load_response_data(file_path)
+
+                responses = []
+
+                # Extract responses based on file structure
+                if isinstance(data, dict):
+                    if "response" in data:
+                        # Single response format
+                        responses.append(data["response"])
+                        logger.info("Found single response format")
+                    elif "generations" in data:
+                        # Multiple generations format
+                        for gen in data["generations"]:
+                            if isinstance(gen, dict) and "response" in gen:
+                                responses.append(gen["response"])
+                        logger.info(f"Found multiple responses format: {len(responses)} responses")
+
+                if not responses:
+                    logger.warning(f"No responses found in {json_file}")
+                    continue
+
+                # Generate image for each response
+                total_responses = len(responses)
+                for idx, response in enumerate(responses, 1):
+                    # Skip empty responses
+                    if not response or not response.strip():
+                        logger.warning(f"Skipping empty response {idx}/{total_responses}")
+                        continue
+
+                    logger.info(f"\nProcessing response {idx}/{total_responses} from {json_file}")
+
+                    # Create prompt from response
+                    prompt = create_prompt_from_response(response)
+                    token_count = count_tokens(prompt)
+                    logger.info(f"Generated prompt ({token_count} tokens): {prompt[:100]}...")
+
+                    # Generate unique output filename
+                    base_name = os.path.splitext(json_file)[0]
+                    if len(responses) > 1:
+                        output_file = os.path.join(output_dir, f"{base_name}_response_{idx}.png")
+                    else:
+                        output_file = os.path.join(output_dir, f"{base_name}.png")
+
+                    logger.info(f"Generating image: {output_file}")
+
+                    # Set negative prompt
+                    negative_prompt = "blurry, low quality, distorted, deformed, disfigured, bad anatomy, ugly, duplicate, error, text, watermark, signature, out of frame, extra fingers, mutated hands, monochrome, grainy"
+
+                    generate_image(
+                        pipe=pipe,
+                        prompt=prompt,
+                        negative_prompt=negative_prompt,
+                        output_file=output_file,
+                        num_steps=30,
+                        guidance_scale=7.5,
+                        size=(512, 512)
+                    )
+
+                    logger.info(f"Completed response {idx}/{total_responses} from {json_file}")
+
+                    # Clear memory between generations
+                    torch.cuda.empty_cache()
+                    gc.collect()
+
+                logger.info(f"Completed processing file {file_idx}/{total_files}: {json_file}\n")
+
+            except Exception as e:
+                logger.error(f"Error processing file {json_file}: {str(e)}")
+                continue
+
+    except Exception as e:
+        logger.error(f"Error in batch generation: {str(e)}")
+        raise
+    finally:
+        clear_gpu_memory()
+
+if __name__ == "__main__":
+    try:
+        # Check dependencies
+        check_dependencies()
+
+        # Configuration
+        response_dir = "../llama7b-chat/actor-method/response_data"
+        output_dir = "outputs/generated_images"
+
+        # Generate images from all JSON files
+        batch_generate_images(response_dir, output_dir)
+
+        print(f"\nImage generation completed successfully!")
+        print(f"Output directory: {output_dir}")
+
+    except Exception as e:
+        logger.error(f"Fatal error: {str(e)}")
+        sys.exit(1)
+    finally:
+        # Final cleanup
+        clear_gpu_memory()
