+++ python/llama7b-chat/actor-method/llm_generation.log
@@ -1,0 +1,48 @@
+2025-03-16 14:12:14,207 - INFO - Attempting 10 different scenarios
+2025-03-16 14:12:14,207 - INFO - Setting up environment
+2025-03-16 14:12:14,257 - INFO - Using GPU: NVIDIA GeForce RTX 4080 SUPER with 15.99 GB memory
+2025-03-16 14:12:14,258 - INFO - Loading model and tokenizer
+2025-03-16 14:12:15,063 - INFO - We will use 90% of the memory on device 0 for storing the model, and 10% for the buffer to avoid OOM. You can set `max_memory` in to a higher value to use more memory (at your own risk).
+2025-03-16 14:12:20,930 - INFO - Generating scenario 1/10
+2025-03-16 14:12:20,930 - INFO - Generating response
+2025-03-16 14:12:32,975 - INFO - Response generated in 12.04 seconds
+2025-03-16 14:12:32,976 - INFO - Saved intermediate results to response-agent_20250316_141232_intermediate.json
+2025-03-16 14:12:34,976 - INFO - Generating scenario 2/10
+2025-03-16 14:12:34,976 - INFO - Generating response
+2025-03-16 14:12:45,338 - INFO - Response generated in 10.36 seconds
+2025-03-16 14:12:45,338 - INFO - Saved intermediate results to response-agent_20250316_141245_intermediate.json
+2025-03-16 14:12:47,340 - INFO - Generating scenario 3/10
+2025-03-16 14:12:47,340 - INFO - Generating response
+2025-03-16 14:12:58,521 - INFO - Response generated in 11.18 seconds
+2025-03-16 14:12:58,522 - INFO - Saved intermediate results to response-agent_20250316_141258_intermediate.json
+2025-03-16 14:13:00,527 - INFO - Generating scenario 4/10
+2025-03-16 14:13:00,527 - INFO - Generating response
+2025-03-16 14:13:11,618 - INFO - Response generated in 11.09 seconds
+2025-03-16 14:13:11,618 - INFO - Saved intermediate results to response-agent_20250316_141311_intermediate.json
+2025-03-16 14:13:13,619 - INFO - Generating scenario 5/10
+2025-03-16 14:13:13,619 - INFO - Generating response
+2025-03-16 14:13:15,058 - INFO - Response generated in 1.44 seconds
+2025-03-16 14:13:15,058 - INFO - Saved intermediate results to response-agent_20250316_141315_intermediate.json
+2025-03-16 14:13:17,058 - INFO - Generating scenario 6/10
+2025-03-16 14:13:17,059 - INFO - Generating response
+2025-03-16 14:13:28,226 - INFO - Response generated in 11.17 seconds
+2025-03-16 14:13:28,226 - INFO - Saved intermediate results to response-agent_20250316_141328_intermediate.json
+2025-03-16 14:13:30,233 - INFO - Generating scenario 7/10
+2025-03-16 14:13:30,233 - INFO - Generating response
+2025-03-16 14:13:41,185 - INFO - Response generated in 10.95 seconds
+2025-03-16 14:13:41,185 - INFO - Saved intermediate results to response-agent_20250316_141341_intermediate.json
+2025-03-16 14:13:43,185 - INFO - Generating scenario 8/10
+2025-03-16 14:13:43,186 - INFO - Generating response
+2025-03-16 14:13:54,045 - INFO - Response generated in 10.86 seconds
+2025-03-16 14:13:54,045 - INFO - Saved intermediate results to response-agent_20250316_141354_intermediate.json
+2025-03-16 14:13:56,050 - INFO - Generating scenario 9/10
+2025-03-16 14:13:56,050 - INFO - Generating response
+2025-03-16 14:14:05,732 - INFO - Response generated in 9.68 seconds
+2025-03-16 14:14:05,733 - INFO - Saved intermediate results to response-agent_20250316_141405_intermediate.json
+2025-03-16 14:14:07,739 - INFO - Generating scenario 10/10
+2025-03-16 14:14:07,739 - INFO - Generating response
+2025-03-16 14:14:18,519 - INFO - Response generated in 10.78 seconds
+2025-03-16 14:14:18,519 - INFO - Saved intermediate results to response-agent_20250316_141418_intermediate.json
+2025-03-16 14:14:18,530 - INFO - Process completed. Successful generations: 10
+2025-03-16 14:14:18,530 - INFO - Failed generations: 0
+2025-03-16 14:14:18,530 - INFO - Final results saved to response-agent_20250316_141418_final.json
