+++ python/llama7b-chat/requirements.txt
@@ -1,0 +1,79 @@
+absl-py==2.1.0
+accelerate==1.5.2
+aiohappyeyeballs==2.6.1
+aiohttp==3.11.13
+aiosignal==1.3.2
+attrs==25.3.0
+bitsandbytes==0.45.3
+certifi==2025.1.31
+charset-normalizer==3.4.1
+cmake==3.31.6
+datasets==2.18.0
+dill==0.3.8
+filelock==3.18.0
+frozenlist==1.5.0
+fsspec==2024.2.0
+grpcio==1.71.0
+huggingface-hub==0.29.3
+idna==3.10
+Jinja2==3.1.6
+Markdown==3.7
+markdown-it-py==3.0.0
+MarkupSafe==3.0.2
+mdurl==0.1.2
+mpmath==1.3.0
+multidict==6.1.0
+multiprocess==0.70.16
+networkx==3.4.2
+ninja==1.11.1.3
+numpy==2.2.3
+nvidia-cublas-cu12==12.4.5.8
+nvidia-cuda-cupti-cu12==12.4.127
+nvidia-cuda-nvrtc-cu12==12.4.127
+nvidia-cuda-runtime-cu12==12.4.127
+nvidia-cudnn-cu12==9.1.0.70
+nvidia-cufft-cu12==11.2.1.3
+nvidia-curand-cu12==10.3.5.147
+nvidia-cusolver-cu12==11.6.1.9
+nvidia-cusparse-cu12==12.3.1.170
+nvidia-cusparselt-cu12==0.6.2
+nvidia-nccl-cu12==2.21.5
+nvidia-nvjitlink-cu12==12.4.127
+nvidia-nvtx-cu12==12.4.127
+packaging==24.2
+pandas==2.2.3
+peft==0.14.0
+pillow==11.1.0
+propcache==0.3.0
+protobuf==6.30.1
+psutil==7.0.0
+pyarrow==19.0.1
+pyarrow-hotfix==0.6
+Pygments==2.19.1
+python-dateutil==2.9.0.post0
+pytz==2025.1
+PyYAML==6.0.2
+regex==2024.11.6
+requests==2.32.3
+rich==13.9.4
+safetensors==0.5.3
+setuptools==76.0.0
+six==1.17.0
+sympy==1.13.1
+tensorboard==2.19.0
+tensorboard-data-server==0.7.2
+tokenizers==0.21.1
+torch==2.6.0
+torchaudio==2.6.0
+torchvision==0.21.0
+tqdm==4.67.1
+transformers==4.49.0
+triton==3.2.0
+trl==0.15.2
+typing_extensions==4.12.2
+tzdata==2025.1
+urllib3==2.3.0
+Werkzeug==3.1.3
+wheel==0.45.1
+xxhash==3.5.0
+yarl==1.18.3
