+++ python/llama7b-chat/test-script.py
@@ -1,0 +1,5 @@
+from huggingface_hub import snapshot_download
+
+# This will download the model to a local cache directory and return the path
+local_dir = snapshot_download("NousResearch/Llama-2-7b-chat-hf")
+print(f"Model downloaded to: {local_dir}")
